requirements.txt
Django
django-libs
python-dateutil
django-currency-history
